app context para migracion de base de datos

+++ DBFrutos.Dominio/Entidades/Usuario.cs
@@ -6,7 +6,7 @@ namespace DBFrutos.Dominio
         public TipoUsuario Tipousuario { get; set; }
         public string Nombre { get; set; }
         public int Documento { get; set; }
-        public bigint Telefono { get; set; }
+        public int Telefono { get; set; }
         public string Direccion { get; set; }
         public string Email { get; set; }
         public string Ciudad { get; set; }
